Matrices inside the shader
The lattice matrices remain inside the shader rather than being sent via push constants. There were a few problems which led to artifacts when using any other code than the one that is currently implemented (still a mystery as to why).
The compiler should optimize the shader accordingly though so that matrix calculation is done once when compiling (I hope).