add calculation of one custommip level
One mip level can now be calculated on CPU with a custom function to blend 4 texels. This will be used to create minmax-mipmaps for later use on GPU.
Note: The resulting mipmap is only in CPU memory for now.
Also updated the noise_heightmap.png to have a color depth of 32