Added interpolation compensation for minmax generation
The minmax calculation on CPU now correctly accounts for bilinear sampling. As a consequence, the minmax range is a little bit bigger now.
Added a button in the ImGUI to switch between raymarching and tessellation-free displacement mapping.
UV to barycentric coordinates matrix is now calculated per triangle on CPU and sent to shader.